memebot_container/requirements.txt
pandas
numpy
faiss-cpu
transformers
flask
onnxruntime